# crates/userserver/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BSCP API",
    "version": "1.0.0",
    "description": "Beorn's Stupid Chat Protocol — user server API. This document is only served by debug builds."
  },
...
  "components": {
    "securitySchemes": {
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "bscp_session" },
      "sessionHeader": { "type": "apiKey", "in": "header", "name": "X-Session-Token" }
    },
    "schemas": {
...
      "Profile": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "example": "alice@localhost:5000" },
          "display_name": { "type": "string" },
          "profile_pic": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["online", "offline", "away", "dnd"] },
          "is_admin": { "type": "boolean" },
          "is_primary_admin": { "type": "boolean" },
          "is_2fa_enabled": { "type": "boolean" },
          "bio": { "type": "string", "nullable": true },
          "storage_limit_mb": { "type": "integer" }
        }
      },
...
      "Webhook": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "url": { "type": "string" },
          "profile_pic": { "type": "string", "nullable": true },
          "created_at": { "type": "number" },
          "last_used": { "type": "number", "nullable": true }
        }
      }
    }
  },
  "paths": {
...
    "/api/users/me": {
      "get": { "tags": ["users"], "summary": "Current user's profile", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Profile" } } } }, "401": { "description": "Unauthenticated" } } },
      "patch": { "tags": ["users"], "summary": "Update display name / bio", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "display_name": { "type": "string" }, "bio": { "type": "string", "nullable": true } } } } } }, "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Profile" } } } } } }
    },
    "/api/users/": {
      "get": { "tags": ["users"], "summary": "List all users (admin)", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Profile" } } } } }, "403": { "description": "Not an admin" } } }
    },
    "/api/users/{full_id}": {
      "get": { "tags": ["users"], "summary": "Look up a profile by username@domain (federation-aware)", "parameters": [{ "name": "full_id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Profile" } } } }, "404": { "description": "Not found" }, "502": { "description": "Remote server unreachable" } } },
      "delete": { "tags": ["users"], "summary": "Deactivate a local user (admin)", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "parameters": [{ "name": "full_id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "OK" } } }
    },
...
    "/api/users/me/push/subscribe": {
      "post": { "tags": ["users"], "summary": "Register a browser push subscription", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "endpoint": { "type": "string" }, "keys": { "type": "object", "properties": { "p256dh": { "type": "string" }, "auth": { "type": "string" } } } } } } } }, "responses": { "200": { "description": "OK" } } },
      "delete": { "tags": ["users"], "summary": "Remove a browser push subscription", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "parameters": [{ "name": "endpoint", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "OK" } } }
    },
    "/api/users/me/activity": {
      "post": { "tags": ["users"], "summary": "Record an activity ping (presence)", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "responses": { "200": { "description": "OK" } } }
    },
    "/api/users/me/picture": {
      "post": { "tags": ["users"], "summary": "Upload a profile picture", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "requestBody": { "required": true, "content": { "multipart/form-data": { "schema": { "type": "object", "properties": { "file": { "type": "string", "format": "binary" } } } } } }, "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "profile_pic": { "type": "string", "nullable": true } } } } } } } },
      "delete": { "tags": ["users"], "summary": "Remove the profile picture", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "responses": { "200": { "description": "OK" } } }
    },
    "/api/users/me/2fa/setup": {
      "post": { "tags": ["users"], "summary": "Begin 2FA enrolment — returns secret + QR", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "secret": { "type": "string" }, "qr_code": { "type": "string" }, "provisioning_uri": { "type": "string" } } } } } } } }
    },
    "/api/users/me/2fa/enable": {
      "post": { "tags": ["users"], "summary": "Confirm and enable 2FA", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["otp"], "properties": { "otp": { "type": "string" } } } } } }, "responses": { "200": { "description": "OK" } } }
    },
    "/api/users/me/2fa/disable": {
      "post": { "tags": ["users"], "summary": "Disable 2FA (password required)", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["password"], "properties": { "password": { "type": "string" } } } } } }, "responses": { "200": { "description": "OK" }, "403": { "description": "Invalid password" } } }
    },
...
    "/api/user/webhooks": {
      "get": { "tags": ["webhooks"], "summary": "List your webhooks", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Webhook" } } } } } } },
      "post": { "tags": ["webhooks"], "summary": "Create a webhook (name is immutable)", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" }, "avatar_url": { "type": "string", "nullable": true } } } } } }, "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Webhook" } } } } } }
    },
    "/api/user/webhooks/{id}": {
      "delete": { "tags": ["webhooks"], "summary": "Delete a webhook", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "204": { "description": "Deleted" } } }
    },
    "/api/user/webhooks/{id}/regenerate": {
      "post": { "tags": ["webhooks"], "summary": "Regenerate a webhook token", "security": [{ "sessionCookie": [] }, { "sessionHeader": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "url": { "type": "string" } } } } } } } }
    },
...
  }
}